partial progress towards new organization
