+++ rtx_package/src/engine/tex_kern.rs
@@ -1,0 +1,8 @@
+//! TeX Kern
+//! 
+//! Core TeX Implementation for LaTeXML
+
+use crate::prelude::*;
+LoadDefinitions!({
+  DefRegister!("\\lastkern" => Dimension::new(0), readonly => true);
+});
